SUKU mad refactoring progress
