+++ src/app/main/grid-layout/grid-modules/overlays/OverlayToggle.svelte
@@ -1,6 +1,6 @@
 <script>
 
-  import { appSettings } from '../../../_stores/app-helper.store.js';
+  import { appSettings } from '../../../runtime/app-helper.store.js';
 
   export let type;
