+++ src/app/main/panels/(unsused) config-library/built-in-configs.js
