+++ src/app/main/panels/(unsused) config-library/config-picker/Folder.svelte
