+++ src/app/main/panels/configuration/components/DynamicWrapper.svelte
@@ -2,10 +2,15 @@
   import { sineOut } from 'svelte/easing';
 
   import { fade } from 'svelte/transition';
+  import { get } from 'svelte/store';
 
   import Options from './Options.svelte';
 
-  import { actionPrefStore, configNodeBinding } from '../../../_stores/app-helper.store.js';
+  import { runtime, user_input, localDefinitions, luadebug_store, appMultiSelect } from '../../../../runtime/runtime.store.js';
+
+  import _utils from '../../../../runtime/_utils';
+
+  import { configNodeBinding } from '../../../../runtime/app-helper.store.js';
 
   export let config = '' //{desc: 'unnamed', rendering: 'standard', id: ''};
   export let configs
@@ -13,7 +18,32 @@
   export let disable_pointer_events = false;
 
   export let toggle = false;
-  const color = () => "rgb(" + Math.floor(Math.random() * 256) + "," + Math.floor(Math.random() * 256) + ","+ Math.floor(Math.random() * 256) + ")";
+
+  function handleConfigChange({configName}){
+
+    // when rendering the Else and End config-blocks, they automatically send out their respective values
+    // this results in config change trigger, which should not be sent out to grid, consider it as AUTO change
+
+    const li = get(user_input);
+
+    const dx = li.brc.dx;
+    const dy = li.brc.dy;
+    const page =  li.event.pagenumber;
+    const element = li.event.elementnumber;
+    const event = li.event.eventtype;
+    const actionstring = _utils.configMerge({config: configs});
+
+    if(configName == 'End' || configName == 'Else'){
+      runtime.update_event_configuration(dx, dy, page, element, event, actionstring, 'EDITOR_EXECUTE');
+
+    } else {
+      runtime.update_event_configuration(dx, dy, page, element, event, actionstring, 'EDITOR_EXECUTE');
+      runtime.send_event_configuration_to_grid(dx, dy, page, element, event);
+    }
+
+    localDefinitions.update(configs);
+
+  }
 
   function heightChange(node, {
     delay = 0,
@@ -32,13 +62,7 @@
     };
   }
 
-  let advancedView = false;
-  // default advanced view behaviour
-  // when the advanced options are open then show the actions with disabled user interactions
-  $: ($actionPrefStore.advanced.index == index && $actionPrefStore.advanced.visible) ? advancedView = true : advancedView = false;
-  // CodeBlock different advanced view behaviour due to different update / render process
-  $: ($actionPrefStore.advanced.index == index && $actionPrefStore.advanced.visible && config.information.name == 'CodeBlock') ? toggle = false : null;
-  
+
 </script>
 
 
@@ -47,60 +71,52 @@
     {#if config.information.rendering == 'standard'}
 
       <carousel 
-          class="flex flex-grow text-white cursor-pointer group"
-          id="cfg-{index}" 
-          movable={config.information.rendering == 'standard' || config.information.name == 'If' ? true : false } 
-          config-component={config.information.name} 
-          config-id={config.id}>
 
-            <div 
-              on:click={()=>{toggle = ! toggle;}}
-              class="{disable_pointer_events ? 'pointer-events-none' : ''} flex relative ">
-              <icon style="background-color:{config.information.color}" class="flex group-hover:bg-opacity-75 items-center p-2">
-                <div class="w-6 h-6">
-                  {@html config.information.icon ? config.information.icon : ' '}         
-                </div>              
-              </icon>
-            </div>
+        class="flex flex-grow text-white cursor-pointer group"
+        id="cfg-{index}" 
+        movable={config.information.rendering == 'standard' || config.information.name == 'If' ? true : false } 
+        config-component={config.information.name} 
+        config-id={config.id}>
 
-            {#if !toggle && !advancedView}
-              <name on:click={()=>{toggle = true;}}  class="pl-4 flex items-center w-full bg-secondary group-hover:bg-select-saturate-10 py-2">
-                <span class="block">{config.information.desc}</span> 
-                <!--<span style="overflow:hidden !important;" class="pl-2 font-mono text-gray-500 inline-block max-w-xs overflow-ellipsis whitespace-nowrap">{config.script}</span>-->
-              </name>
-            {/if}
+          <div 
+            on:click={()=>{toggle = ! toggle;}}
+            class="{disable_pointer_events ? 'pointer-events-none' : ''} flex relative ">
+            <icon style="background-color:{config.information.color}" class="flex group-hover:bg-opacity-75 items-center p-2">
+              <div class="w-6 h-6">
+                {@html config.information.icon ? config.information.icon : ' '}         
+              </div>              
+            </icon>
+          </div>
+
+          {#if !toggle}
+            <name on:click={()=>{toggle = true;}}  class="pl-4 flex items-center w-full bg-secondary group-hover:bg-select-saturate-10 py-2">
+              <span class="block">{config.information.desc}</span>
+            </name>
+          {/if}
       
       </carousel>
           
-      {#if toggle || advancedView}
+      {#if toggle}
 
-      <container in:heightChange class="{advancedView ? 'opacity-50 pointer-events-none' : ''} w-full flex bg-secondary bg-opacity-25 rounded-b-lg">
-        <fader-transition class="w-full" in:fade={{delay: 200}} out:fade={{delay:0,duration:0}}>
+        <container in:heightChange class=" w-full flex bg-secondary bg-opacity-25 rounded-b-lg">
+          <fader-transition class="w-full" in:fade={{delay: 200}} out:fade={{delay:0,duration:0}}>
 
-          <svelte:component 
-          this={config.component} 
-          {index} 
-          {config}
-          slot="humanify"
-          on:output={(e)=>{config.script = e.detail.script; handleConfigChange({configName: config.information.name}); configs = configs;}}/>
+            <svelte:component 
+              this={config.component} 
+              {index} {config}
+              on:output={(e)=>{config.script = e.detail.script; handleConfigChange({configName: config.information.name}); configs = configs;}}
+            />
 
-        </fader-transition>
-      </container>
-<!--
-      {:else if advancedView}
+          </fader-transition>
+        </container>
 
-      <name class="{advancedView ? 'opacity-50 pointer-events-none' : ''} pl-4 flex items-center w-full bg-secondary text-white group-hover:bg-select-saturate-10 py-2">
-        <span class="block">{config.information.desc}</span> 
-        <span style="overflow:hidden !important;" class="pl-2 font-mono text-gray-500 inline-block max-w-xs overflow-ellipsis whitespace-nowrap">{config.script}</span>
-      </name>
--->
       {/if}
       
     {:else if config.information.rendering == 'modifier'}
 
       <div 
         id="cfg-{index}" 
-        movable={config.information.rendering == 'standard' || config.information.name == 'If' ? true : false } 
+        movable={config.information.name == 'If'} 
         config-component={config.information.name} 
         config-id={config.id}
         class="flex w-full flex-col">
@@ -108,53 +124,18 @@
         <div class="{disable_pointer_events ? 'pointer-events-none' : ''} w-full flex relative">
 
           <svelte:component 
-          this={config.component} 
-          {index} 
-          {config}
-          slot="humanify"
-          on:output={(e)=>{config.script = e.detail.script; handleConfigChange({configName: config.information.name}); configs = configs;}}/>
+            this={config.component} 
+            {index} {config}
+            on:output={(e)=>{config.script = e.detail.script; handleConfigChange({configName: config.information.name}); configs = configs;}}
+          />
 
-        </div>
-
-      </div>
-
-      {:else if config.information.rendering == 'fixed'}
-
-      <div 
-        id="cfg-{index}" 
-        movable={true} 
-        config-component={config.information.name} 
-        config-id={config.id}
-        class="flex w-full flex-grow">
-
-        <div 
-          class="{disable_pointer_events ? 'pointer-events-none' : ''} flex relative ">
-          <icon style="background-color:{config.information.color}" class="flex group-hover:bg-opacity-75 items-center p-2">
-            <div class="w-6 h-6">
-              {@html config.information.icon ? config.information.icon : ' '}         
-            </div>              
-          </icon>
-        </div>
-
-        <div class="{disable_pointer_events ? 'pointer-events-none' : ''} w-full flex relative">
-
-          <svelte:component 
-          this={config.component} 
-          {index} 
-          {config}
-          slot="humanify"
-          on:output={(e)=>{config.script = e.detail.script; handleConfigChange({configName: config.information.name}); configs = configs;}}/>
-       
         </div>
 
       </div>
 
     {/if}
 
-
     <Options {toggle} {index} {configs} rendering={config.information.rendering} componentName={config.information.name} />
-
-
 
 </wrapper>
