+++ src/app/main/panels/configuration/components/Pages.svelte
@@ -4,9 +4,9 @@
 
   import { get } from 'svelte/store';
 
-  import { logger, unsaved_changes, runtime } from '../../../runtime/runtime.store.js';
+  import { logger, unsaved_changes, runtime } from '../../../../runtime/runtime.store.js';
 
-  import TooltipSetter from '../../user-interface/tooltip/TooltipSetter.svelte';
+  import TooltipSetter from '../../../user-interface/tooltip/TooltipSetter.svelte';
 
   export let pages;
