--- src/app/shared/helpers/global-helper.js
@@ -1,5 +1,0 @@
-const shell = require('electron').shell
-
-export function openInBrowser(url){
-  shell.openExternal(url)
-}
